[Fix #26] reconf strategies must prepare agents for reconfiguration
When an agent is reconfigured, it must perform some cleanup tasks:
it removes all resource requests for the reconfigured task and aborts
execution of its current role if that role belongs to that task.

Previously, this was only done on the agent initiating reconfiguration.
This is incorrect, it must be done on all affected agents.